+++ http-cache-tower-server/CHANGELOG.md
@@ -1,5 +1,20 @@
 # Changelog
 
+## [0.2.4] - 2026-05-29
+
+### Added
+
+- `manager-redb` feature flag for `RedbManager` support
+
+### Changed
+
+- Updated `http-cache` dependency to 1.0.0-alpha.7
+- Default feature changed from `manager-cacache` to `manager-redb`; enable `manager-cacache` explicitly to keep using `CACacheManager`
+
+### Fixed
+
+- `tokio` dependency now enables the `rt` feature required for `tokio::spawn` (previously satisfied only transitively via the `manager-cacache` feature)
+
 ## [0.2.3] - 2026-04-17
 
 ### Changed